Revert "atomic design 적용"
