--- frontend/.storybook/main.js
@@ -1,20 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
-  addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
-  webpackFinal: async (config, { configType }) => {
-    config.resolve.alias = {
-      ...config.resolve.alias,
-      '@components': path.resolve(__dirname, '../src/components'),
-      '@atoms': path.resolve(__dirname, '../src/components/atoms'),
-      '@molecules': path.resolve(__dirname, '../src/components/molecules'),
-      '@organisms': path.resolve(__dirname, '../src/components/organisms'),
-      '@templates': path.resolve(__dirname, '../src/components/templates'),
-      '@pages': path.resolve(__dirname, '../src/components/pages'),
-      '@utils': path.resolve(__dirname, '../src/utils'),
-      '@styles': path.resolve(__dirname, '../src/styles'),
-    };
-    return config;
-  },
-};
