--- frontend/.storybook/preview.js
@@ -1,4 +1,0 @@
-
-export const parameters = {
-  actions: { argTypesRegex: "^on[A-Z].*" },
-}
