--- frontend/src/components/atoms/Button.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const Button = styled.button`
-  width: 100%;
-  height: 33px;
-  font-weight: bold;
-  color: #ffffff;
-  background-color: ${(props) => props.backgroundColor};
-  border-radius: 3px;
-  border: none;
-  outline: none;
-  cursor: pointer;
-`;
-
-export default Button;
