--- frontend/src/components/atoms/Button.stories.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import Button from './Button';
-
-export default {
-  title: 'atoms/Button',
-  component: Button,
-  argTypes: {
-    backgroundColor: { control: 'color' },
-  },
-};
-
-const Template = (args) => <Button {...args} />;
-
-export const Primary = Template.bind({});
-Primary.args = {};
