--- frontend/src/components/atoms/CheckBox.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const CheckBox = (props) => {
-  return <input type="checkbox" />;
-};
-export default CheckBox;
