--- frontend/src/components/atoms/FlexItem.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const FlexItem = styled.div`
-  display: flex;
-  justify-content: center;
-`;
-export default FlexItem;
