--- frontend/src/components/atoms/Input.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const Input = styled.input`
-  width: 250px;
-  height: 25px;
-  border: 1px solid #d6d8db;
-`;
-
-export default Input;
