--- frontend/src/components/atoms/Label.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const Label = styled.p`
-  margin-bottom: 5px;
-  font-weight: bold;
-  font-size: 15px;
-`;
-
-export default Label;
