--- frontend/src/components/atoms/Text.stories.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import Text from '@atoms/Text';
-
-export default {
-  title: 'atoms/Text',
-  component: Text,
-};
-
-const Template = (args) => <Text {...args} />;
-
-export const Primary = Template.bind({});
-Primary.args = {};
