--- frontend/src/components/atoms/TextButton.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const TextButton = styled.input`
-  margin: 15px 5px 20px 5px;
-  background-color: #ffffff;
-  color: #397eff;
-  border: none;
-  outline: none;
-  cursor: pointer;
-  font-size: 14px;
-  font-weight: bold;
-`;
-
-export default TextButton;
