+++ frontend/src/components/common/FlexBox.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import styled from "styled-components";
+
+const FlexBox = styled.div`
+display:flex;
+flex-direction:${(props)=>props.direction || "row"};
+`;
+export default FlexBox;
