+++ frontend/src/components/common/FlexItem.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import styled from "styled-components";
+
+const FlexItem = styled.div`
+display:flex;
+flex-grow:1;
+justify-content: center;
+`;
+export default FlexItem;
