--- frontend/src/components/molecules/InputBox.stories.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import InputBox from '@molecules/InputBox';
-
-export default {
-  title: 'molecules/InputBox',
-  component: InputBox,
-};
-
-const Template = (args) => <InputBox {...args} />;
-
-export const Primary = Template.bind({});
-Primary.args = {
-  label: 'test',
-  type: 'text',
-};
