--- frontend/src/components/organisms/IssueList.stories.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import IssueList from '@organisms/IssueList';
-
-export default {
-  title: 'organisms/IssueList',
-  component: IssueList,
-  argTypes: {},
-};
-
-const Template = (args) => <IssueList />;
-
-export const Primary = Template.bind({});
-Primary.args = {};
