--- frontend/src/components/pages/LoginPage.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import LoginPageTemplate from '@templates/LoginPageTemplate';
-import '@styles/LoginPage.css';
-
-const LoginPage = () => {
-  return <LoginPageTemplate />;
-};
-
-export default LoginPage;
